tree: reimplement range_delete with message insertion
The naive way to delete a key range is to insert a bunch of deletion
messages into the tree, then wait for them to flush down until the pairs
are actually deleted. This is very suboptimal, because large ranges can
result in very large amounts of insertions.

Felix Wiedemann optimised this cleverly, deleting entire subtrees
without message insertion, but didn't implement the subsequent
rebalancing/merge operation. I have doubts that an ad-hoc implementation
of mine would cover all edge-cases soundly, and don't have the time to
properly complete this optimisation.

If sufficient time is available, it should definitely be reimplemented.