buffer: replace with block-aligned buffer variants
The previous buffers did not align to page boundaries, which we'll need
for direct IO, and relied on careful state management by the user to
avoid leaks. It also didn't allow for buffer growth.

With current stable std, there's no way to replace the allocator of a
Vec locally, but once that is stabilised, AlignedStorage could be
rewritten as an allocator for a Vec.