storage_pool: support multiple classes, and reduce queue delay
The old queueing mechanism had no way to mark a future as completed,
instead it had to fall back to completing random futures until it found
the one it wanted to actually remove.

This is now possible, and hopefully this reduces the need for enqueue to
apply backpressure, by keeping the queue free of completed futures.