default_message_action: zero-copy upserts
Deserialisation and serialisation of DefaultMessageActions
dominated CPU profiles, occasionally larger than 70% of CPU time.

The deserialisation of Upserts was particularly expensive, having to
allocate and fill multiple Vec, only to append and serialise them
moments later.

Now we avoid having to ever deserialise an upsert collection into
multiple Vecs, by iterating over the encoded form in-place,
and are able to append to the end without touching the previous ones.