tree: undo pref caching in internal nodes
Internal nodes receive no notification if the node_pointers of their
child buffers are updated, and thus can't cache that information.

Child buffers now only track the accumulated preference of their
messages, and always integrate the node_pointer into the preference
directly. This avoids the need to invalidate a child buffer (or internal
node) when the node_pointer is updated (to a possibly lower class).