cache: add unsafe verification feature
Something was wrong with the size tracking, which caused the tracked cache size
to increase without actually filling up the cache. This resulted in the
cache always triggering evictions, even if it contained only one entry,
Evictions which would eventually trigger write-backs, which trigger
copy-on-write, which triggers addition to the delayed_messages buffer,
which triggers more cache insertions and more writes, thus runaway
writes until all storage is consumed.

With no way of determining where and when the cache size tracking goes
wrong, the only sensible approach is to verify cache state as often as
possible, to pinpoint the code that wrongly adjusts (or forgets to
adjust) cache size.

But the cache uses a HashMap<_, Arc<S>>, where S contains a
user-provided type. To calculate the size of S for verification, we
could add a "Size" bound, but the only actual instantiation of
ClockCache happens with S containing a RwLock<_>. And how do you get the
size of RwLock without risk of deadlocks? With RwLock::get_mut and then
SizeMut::size, but that requires a &mut RwLock<_>, which you can't get out of an Arc<S>.

Thus, because I don't want to rewrite the entire cache system, and
everything that uses it (the entire tree layer, and data_management),
we resort to a bit of completely wrong unsafe code, that just so happens
to work. For now.

Turning an Arc<T> into a &mut T is utterly wrong, and the compiler will
probably feel free to add all sorts of undefined behaviour if
optimisations are on. Hopefully, it will refrain from making it any worse
than it already is, if you keep optimisations turned off.

It has also been feature-gated with a non-existent feature, so that
nobody accidentally uses it, but kept in for desperate future debugging.

To summarise: if you want to use this, add and enable a feature
"cache-paranoia", turn off all optimisations or suffer nasal demons,
and litter the suspicious code with calls to "verify()". Good luck.