Add manual migration functionality (#2)
* Add todo for platform dependent fix

* Allow easier allocation investigation

Add a few tracing and info message at vital points of syncing processes.
We already found a busy lock so there are probably quite a few left on
some occassions..

* Add interface for migrating data (Draft implementation)

We can improve this implementation by not reinserting every single
data which has to be moved.

There are multiple ways to do this:
1. For now the implemented approach is very simple and
reinserts keys as they are requested. Though this instantiates copies
in some cases. (Current)
2. Add a new messages which instantiates the move.
3. Move data with a new interface for leaf nodes which intitiate a
   new writeback, dealing properly with the ObjectHandle life cycle.

* Add tests for writing behavior

Most of these aim to fill the file system to expose unwanted behavior.
Some of these cases require timeouts

* Add migration integration tests

* object: Implement basic migrate

* data_management: Add space accounting for dmu allocations

* tracing: add tracing messages for storage pool

* tests: switch up & down semantics

* dmu/handler: add free tier space func

* buffer: lower growth factor for better utilization

TODO: This leads to slower performance on enlarging of buffers
we might want to repeat experiments with this growth number or
make it configurable. Depending on the context.

* data_management: fix deadlock on failed syncs

The sync was kept in an errnoeous state after it is failing to allocate
enough space in the storage, which lead to deadlocks when dropping the
database object for example.

This commit returns the MID into it's original state and simply will
lead future sync to fail the same way as the first one.
To fix the problem from a user perspective the space accounting should
be expanded to fail early on inserts which exceeded the recorded
capacity.

* data_management: add free space call for database

Ease of use call to retrieve storage information about all storage
tiers.

* tests: add differentiated test for filling and overfilling

* tests: remove name req for flaky test

* tests: increase buffer for invalid migrate

* bectl: update usage of get free space

* rustfmt: format code

* benchmark: fix compilation issues

The benchmarks aren't really that useful anymore and seem abandoned we
might extend them in the future or simply throw them out.

* tests: add readme

* dataset: Use messages for migration

This currently has the down side that the data is only migrated once it
reaches the leafs which depending on the structure of the tree might
take some time. This can be solved by implementing a force flush on
migrations so that all relevant data reaches the leafs on insert.
Though this may require in the end a new message type, which extends the
upsert one.

* docs: add free space tier documentation

* fmt: cargo fmt

* object: use space accounting to pre-check migration

This is an example implemenation on how to use space accounting, though
only migration and only the object store uses it now.

Also we assume the maximum possible size right now, which we might want
to change if more efficient compression is used.

* project: update authors & version

* tests: Change overfull write test to expect error early

* betree: remove unused imports

* betree: handle unused results

* object: add function for implicit up and down migration

* object: simplify migration error

* betree: cargo fmt

* tests: remove write snapshots

* git: add Cargo.lock to gitignore

Since this repository contains a library we should add the lock file to
the ignored files