tests: handle file backed config with locking
This allows for multiple tests to rely on the same files without
additional checking or depedencies between tests. Problematic add the
moment is that some tests are order dependent due to the storage not
being nulled. Reading some data results in _some_ cases in a checksum
error. This happens on either file mode, indicating some problem
which has not been discovered yet.