dataset: add wrapper type to allow dataset share
While the previous implementation was fine to use with async executors
given lifetime conditions are met, which it is still is, we might
require access from less clear situations or threads which outlive it's
purpose of accessing this dataset. For this purpose a shared rwlocked
option is contained in the wrapper, which can be freely shared with all
participants, while after the main owner (owner of the database) decides
to close the dataset all connections are terminated and the wrapper
closed indefinetly.