betree/tests: indicate expected failures
Some tests have been implemented to show that not all situations are solved optimally at the moment, and fail in the current state of the storage stack. Due to the long-term nature of solutions to resolve them, we indicate that they are expected to fail right now, so that the CI may properly interpret when new tests are failing.