data_management: Fixate ObjectPointer Info & Generation
In the effort of improvement of ergonomics for a better developing experience
we fixate some generics which only have a single useful implementor or a single
trait implementor. This is the first commit in this series which focuses on the
ObjectPointer a construct pointing to a disk location of serialized written
data. In discussion within the group we decided that in the future we will have
no sensible use for other types in this position. Also, considering that a
major part of the database implementation is dependent on a certain Info =
DatasetId it is unlikely and impractical to actually exchange this type here.
So we might as well use it as a newtype as is done now.