data_management: clarify ObjectRef v. ObjectRef
These have caused some identification issues for the *human* reader for some
time.  While still being discernible by the omission or specification of the
generic type <P>, the similarity between names especially in the documentation
was annoying at best and a hinderance at worst. This commit atleast resolves
the naming collision which was present by using a fully written name for the
trait and an abbreviation type for the actual implementor.

Also, the existence for the trait has been mentioned in the comment. As we
only use ever one implementor for this trait (The ObjRef (see how
understandable the diff is now) has multiple version it can exists as, so even
if we want to use another serialized representation, e.g. NVRAM, we may simply
interject in this representation there as it is much clearer anyway for all
purposes imaginable *with* the current DMU, for which the traits anyway are
tailored to), normally single-implementor traits are considered bad practice,
but the usage of associated types allows our generics to be a bit better
readable, so we might use them here. For now.