data_management: Remove DmlWithCache trait
This trait was anyway a disnomer, as the "general" Dml trait included cache
dependend information like `CacheValueRef`. A Dml which never uses any cache
also is unlikely and definetly for our purposes not usable, more likely again
that we want specific subtypes that don't use a cache which is better portrayed
as modification to the `ObjectPointer` or `ObjRef`.