data_management: Move DmlWithSpl to Dml
We will not have a Dmu without any storage backend. Even for testing we might
add a dummy backend rather than omitting it completely as this requires
internal control of the Dmu, as the Database requires a storage backend.
Therefore, this has been integrated into the Dml trait to reflect this reality.