dml: move ObjRef From<ObjPtr> impl to own method to avoid misuse
We silently assumed that only root ObjPtrs will be passed to the from/into
calls which might not be clear in each position where we want to use this
transformation. A new method `root_ref_from_obj_ptr` has been added to replace
the original `From` impl which makes this clearly visible while reading.