database: Remove DatabaseBuilder trait
The main purpose of this trait was the "abstracted" construction of a database
from a database configuration. But with how specified the trait has become, and
with how purpose built the Database and interlinking structure were on this
configuration it is unlikely we ever use a *completely* different configuration
in comparison to modifying, extending or shrinking the current structure. Even
less so when we consider dynamic usage of the trait. Therefore this commit
removes `DatabaseBuilder`.