fio: deactivate queue depth translation
Setting the queue depth to 1 (as is done in sync workflows) provokes a deadlock
in haura atm. As some nodes can't get sync'ed, this requires some investigation
but should ideally just exit with an error in this case. To avoid this error in
fio for the moment, and as technically another queue depth is semantically
meant by this in fio, we deactivate this option in this commit.