checksum: add FxHash and GxHash based checksum (#58)
* checksum: add fxhash

This is just to test out the performance compared to the xxhash we've been
using until now.  Early measurements with 4M blobs have shown that it could be
worth experimenting with fxhash which is used in the rust compiler.

* checksum: structure module

* checksum: fix typo

* storage_pool: remove obsolete in_memory module

Noticed this while grepping for XxHash, seemed to have evaded the cleaning
process some generations ago.

* superblock: remove dependency on XxHash

* checksum: add gxhash

This commit required modifying the build context to allow for the AES
optimizations of GxHash. It should not prove to be an issue on the system we
use (x86-64 and maybe ARM64) which I've tested before this commit.

* checksum: remove potential ambiguous module ref

* betree: move .cargo config to workspace level

* checksum: correct module description

* database: make GxHash the default checksum