buffer: fix creeping nulls
This happened in the CI and on the most recent rust compiler on my system in in
debug mode. Previously a lot of .is_null()'s were distributed over the code
this commit cleans this up and by-chance fixed whatever bug in the code flow
was occurring in this case.