tree: fix sequential insertion tree construction
So for quite some time sequential insertion constructed a tree which did not
really adhere to the bepsilon-tree rules. This was due to the nodes-in-cache
optimization in the insertion code which skips insertion into nodes when their
child nodes are in cache. This lead to the case that on sequence many leaves
where created and all the pivots are inserted into the parent node of the last
node in cache, this was never checked bc we only call rebalance on the final
node which was the last node in cache. Now bc of this these parents grew
without checks and pivots were essentially just glued together. First, this
slows down searching in the node. Second, all access guarantees and buffer
spaces normally allowed in the bepsilon tree are gone and with only pivots our
tree essentially behaved like a btree in these scenarios. Why this was never
caught before i don't know but this commit fixes this behavior doing two
things: 1. The `is_too_large` of the node objects now include this space
devision of at maximum B^epsilon space for pivots. Meaning as soon as nodes
overstep this boundary they are split to adhere to bepsilon-tree construction
but might be smaller than 4m, 1m, whatever. This has implication on performance
(positive and negative) but is the correct thing to do. 2. Before we check if
the child of the current node is in cache and can be modified we check if the
current node is already too large if this is the case we DO NOT SKIP THE
CURRENT NODE but instead insert the message into the current internal node.
This causes more operations on insertion but also makes future updates as cheap
as they are actually expected to be with the complexity of the bepsi tree.

In the context of this: Another bug was fixed which highlights how problematic
this behavior was, the `get_with_info` code of the node was not able to fetch
an entry when it was not present in the leaves. Due to the bug when
constructing the tree sequentially this was not caught somehow before. It is
fixed now.